### 2021-11-30 #492 @cboettig
gbifdb: Local Database Interface to 'GBIF'
Submitting Author Name: Carl Boettiger
Submitting Author Github Handle: @cboettig
Repository:  https://github.com/cboettig/gbifdb
Version submitted: 0.1.0
Submission type: Standard
Version accepted: TBD

---

```
Package: gbifdb
Version: 0.1.0
Title: Local Database Interface to 'GBIF'
Description: A high performance interface to the Global Biodiversity
  Information Facility, 'GBIF'.  In contrast to 'rgbif', which can
  access small subsets of 'GBIF' data through web-based queries to
  a central server, 'gbifdb' provides enhanced performance for R users
  performing large-scale analyses on servers and cloud computing
  providers, providing full support for arbitrary 'SQL' or 'dplyr'
  operations on the complete 'GBIF' data tables (now over 1 billion
  records, and over a terrabyte in size.) 'gbifdb' accesses a copy
  of the 'GBIF' data in 'parquet' format, which is already readily
  available in commercial computing clouds such as the Amazon Open
  Data portal and the Microsoft Planetary Computer, or can be
  accessed directly or downloaded
  to any server with suitable bandwidth and storage space.
            person("Carl", "Boettiger", , "cboettig@gmail.com", c("aut", "cre"),
                   comment = c(ORCID = "0000-0002-1642-628X"))
            )
    duckdb (>= 0.2.9),
    DBI
    dbplyr,
    covr,
    aws.s3,
    arrow

```

This package is similar in scope to `rgbif`, but allows direct access to the full occurrence table using `dplyr` functions with or without downloading.

Ecologists etc using / analyzing GBIF data.  In particular, anyone attempting analyses which need the whole GBIF data corpus, like the example shown in the README, cannot easily do so with the standard API / rgbif approach.

As noted above, this is similar in scope to rgbif, but more general and faster.
