+++ tests/testthat/test_gbifdb.R
@@ -18,9 +18,25 @@ test_that("gbif_conn()", {
 
 })
 
+test_that("gbif_dir()", {
+
+  dir <- gbif_dir()
+  expect_true(TRUE)
+})
+
 test_that("gbif_remote()", {
-  skip_on_cran()
+  # skip_on_cran()
   skip_if_offline()
   conn <- gbif_remote(to_duckdb = FALSE)
   expect_true(inherits(conn, "Dataset"))
+})
+
+test_that("gbif_remote(to_duckdb=TRUE)....slow!", {
+  # skip_on_cran()
+  skip_if_offline()
+  library(arrow)
+  library(dplyr)
+  conn <- gbif_remote(to_duckdb = TRUE)
+  expect_true(inherits(conn, "tbl_dbi"))
+  duckdb::dbDisconnect(conn$src$con, shutdown = TRUE)
 })
