+++ man/gbif_dir.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gbif_download.R
+% Please edit documentation in R/gbif_dir.R
 \name{gbif_dir}
 \alias{gbif_dir}
 \title{Default storage location}
@@ -9,4 +9,7 @@ gbif_dir()
 \description{
 Default location can be set with the env var GBIF_HOME,
 otherwise will use the default provided by \code{\link[tools:userdir]{tools::R_user_dir()}}
+}
+\examples{
+gbif_dir()
 }
