### 2021-12-02 #492 @mpadge
@cboettig This is the first submission with an `arrow` dependency, for which standard installation of system libraries are inadequate. I'll rebuild the check system with a pre-install of arrow, and your package will at least pass R CMD check. Thanks for the good test case!
