### 2021-12-02 #492 @cboettig
Thanks @mpadge , but it's good to discover these errors now!  I forgot that `arrow` could optionally be built without S3 support.  I've added checks so the test will skip if `arrow::arrow_info()` reports that we arrow was built without s3 capability.  (No doubt CRAN has at least one machine where that is also the case.  Realistically I will need to skip that check on CRAN too I think; the operation is too network-intensive to be robust and again I don't think there's much value in creating a mock test that passes even when the network is unavailable).
