### 2021-12-03 #492 @jhollist
Hi,@cboettig!  Looking forward to working on this with you!  I am full up today, but wanted to let you know that I have this on my radar and will start working on it early next week.
