### 2021-12-06 #492 @jhollist
@mpadge Have you had a chance to add arrow to the check system?  I want to hold off running the package checks until I know that is ready to roll.
