### 2021-12-06 #492 @mpadge
Yep @jhollist it's up and rolling with arrow pre-compiled for S3 support. You may check away ...
