### 2021-12-06 #492 @cboettig
@jhollist Thanks much!

I've added a `contributing.md` (`usethis::use_tidy_contributing()`), apologies for missing that one.

As mentioned above, I don't think it makes sense to add CITATION at this stage, (R will already do the best option via `citation()`, citeAs will recognize the codemeta.json  (and I think the DESCRIPTION), Zenodo will I think continue to ignore all these and go by the github metadata or it's own .zenodo.json override instead).   I love metadata but I'm not convinced data entry is improved by so much duplication at this stage....  If this ever has an external doc or paper to cite, I'll certainly add that to CITATION at that time....

Also not sure I can do much about code coverage without resorting to silly hacks.  (I totally appreciate the spirit of the rule, but am never fond of putting to much emphasis on such statistics which become relatively easy to hack which only further erodes their utility....)

I think we're clear on the other checks? Cool auto-check system!  Let me know what's next!
