### 2021-12-06 #492 @jhollist
@cboettig A couple of items:

1. Go ahead and still add the CITATION file (see: https://devguide.ropensci.org/building.html?q=citatio#citation-file).  Having this in there will allow you to track any Zenodo DOIs that get minted (although not necessary right now).  Also, it might be useful to provide a citation for the GBIF data/AWS service.

2. I agree with you on the testing.  I think you have a fairly complete test suite.  I will probably ask the reviewers to think about this one as well, just to make sure we aren't missing something.  That being said, nothing to do right now.  I am happy to send on to reviewers with the current test suite.

3.  I can't get the dev version of duckdb installed, but looks like the version on CRAN is > what you require in the DESCRIPTION.

4. When I run `gbif_remote()` I get
```
Error: IOError: When getting information for key 'occurrence/2021-11-01/occurrence.parquet' in bucket 'gbif-open-data-af-south-1': AWS Error [code 15]: No response body.
```
I tried with different version and different bucket but did not work.  Is this my machine (session info below) or do you see the same thing?

So next steps are to get the CITATION, double check that user needs the dev version of duckdb or if CRAN is fine, make sure gbif_remote() is working.    After that I have some other local checks to do then I can pass it on to reviewers.

Make sense?

My Session Info:
```
R version 4.1.0 (2021-05-18)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 18363)

Matrix products: default

locale:
[1] LC_COLLATE=English_United States.1252  LC_CTYPE=English_United States.1252
[3] LC_MONETARY=English_United States.1252 LC_NUMERIC=C
[5] LC_TIME=English_United States.1252

attached base packages:

other attached packages:
[1] dplyr_1.0.7  gbifdb_0.1.0

 [1] arrow_6.0.1      fansi_0.5.0      assertthat_0.2.1 utf8_1.2.2       crayon_1.4.2     R6_2.5.1
 [7] DBI_1.1.1        lifecycle_1.0.1  magrittr_2.0.1   pillar_1.6.4     rlang_0.4.12     remotes_2.4.1
[13] vctrs_0.3.8      generics_0.1.1   ellipsis_0.3.2   tools_4.1.0      bit64_4.0.5      glue_1.5.1
[19] bit_4.0.4        purrr_0.3.4      compiler_4.1.0   pkgconfig_2.0.3  tidyselect_1.1.1 tibble_3.1.6
```
