### 2021-12-06 #492 @jhollist
I can confirm that it is me, not you!  Something with my work machine.  I checked your actions after I sent my response and saw that they worked...  I was able to get it to run on my son's laptop!  I will sequester his machine and do the additional checks.  I'll try to get through that tomorrow.

I'll also dig in a bit more on my machine and see if I can get it working.    Stay tuned!
