### 2021-12-07 #492 @jhollist
@cboettig I figured it out.  I had an existing aws credentials file that I think had expired.  I think authentication was failing and that was causing problems.  Removed my old credentials file and tried again.  Everything working!  Will finish with checks on this tomorrow.
