### 2021-12-07 #492 @cboettig
@jhollist brilliant detective work! :detective: I had forgotten about credentials! actually I'd seen that same problem before with aws.s3 download -- which fails even with valid credentials I believe because this is a public bucket, you can see I override them like so:
https://github.com/cboettig/gbifdb/blob/main/R/gbif_download.R#L33

(not sure if that's the best option really, credentials in `~/.aws` dir could also get in the way I think)  Anyway, would be a great issue to hash out with reviewer input!  Meanwhile I can at least add the issue to the function documentation.
