+++ R/gbif_remote.R
@@ -25,6 +25,7 @@ gbif_remote <-
             stop("please install arrow first")
         }
 
+        unset_aws_env()
         server <-
          arrow::s3_bucket(bucket, ...)
         prefix <- 
@@ -33,6 +34,8 @@ gbif_remote <-
                  "/occurrence.parquet/")
         path <- server$path(prefix)
         df <- arrow::open_dataset(path)
+        
+        ## Consider leaving this to the user to call.
         if (to_duckdb) {
             if (!requireNamespace("dplyr", quietly = TRUE)) {
                 stop("please install dplyr first")
@@ -41,3 +44,18 @@ gbif_remote <-
         }
         df
     }
+
+
+unset_aws_env <- function(){
+    
+    ## Consider re-setting these afterwards.
+    ## What about ~/.aws ?
+    ## Maybe set these to empty strings instead of unsetting?
+    
+    ## Would be nice if we could simply override the detection of these
+    Sys.unsetenv("AWS_DEFAULT_REGION")
+    Sys.unsetenv("AWS_S3_ENDPOINT")
+    Sys.unsetenv("AWS_ACCESS_KEY_ID")
+    Sys.unsetenv("AWS_SECRET_ACCESS_KEY")
+}
+
