### 2021-12-08 #492 @jhollist
---

- As you point out in your Package Description, there is overlap with `rgbif`.  You make the case that the difference is with generality and performance.  I am not too worried about this but I will ask the reviewers to think about it and would ask you to think a bit more on what distinguishes `gbifdb` from `rgbif`.  At a minimum, the case you make in the DESCRIPTION, would be useful to have on the README as well.
- Link to code of conduct in CONTRIBUTING.md is busted
- As we discussed, your current test suite is fine so you can ignore that `X`!

I will start working on finding reviewers.  Looking forward to working on this!

---
