### 2022-01-12 #492 @stephhazlitt
## {gbifdb} Package Review

Thank you for the opportunity to test drive and review this great new R package. {gbifdb} brings together many things near and dear to me -- easier access to lots and lots of open biodiversity data 🦋🦉🦡, R,  and leveraging new open source tools like Arrow 🏹 and DuckDB 🦆 to make all of this faster and possible on a "regular" laptop (for my review I used a MacBook Pro M1 2020). I hope this feedback is helpful, please feel free to reach out if there are follow up questions or if there is anything I can do to support moving forward.

- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work.

Estimated hours spent reviewing: 5ish

---

#### Remote Data Access

- I was able to successfully pull some GBIF data using `gbif_remote()` out of the box 🎉 :

```
library(gbifdb)
library(tictoc)

gbif_s3 <- gbif_remote()

tic()
df <- gbif_s3 |>
  filter(species == "Haematopus bachmani",
         year > 2018,
         stateprovince == "British Columbia") |>
  count(year) |>
  collect()
toc()

## Error: IOError: AWS Error [code 99]: curlCode: 28, Timeout was reached
## 6098.66 sec elapsed
```

- I was initially surprised how long it took though, a bit over an hour and a half, but then I came around to that being probably expected given my not great 50:20 copper internet plan. I think letting users know roughly what to expect re: compute time would be useful for managing expectations and helping make workflow decisions (to download or not to download).
- I also encountered the curl error above, which I assumed was gremlins with a remote workflow so I just tried again and voila. It might be worth proactively documenting this potential outcome with some try-try-again troubleshooting advice?
- I spent some time on the GBIF website and used the example data and output from the `colnames(gbif)` in the local data section of the README to get familiar with the GBIF data first. I think it would be useful to provide new (GBIF) users with the full name, the url and a couple more sentences on GBIF, what data is available (and what is not) and the easiest way to access metadata.
- A possible future enhancement might be some helper functions to get this metadata into the R environment (column names, data dictionaries, even if stored in the package?) to help users filter and summarise efficiently

#### Local Data Access

- I started to download the GBIF data set to run over night, only to find that my laptop went to sleep on me in the wee hours which borked the process. Maybe some user tips for getting a machine ready for a _big_ download?
- As a time-saver I only downloaded about 24 parquet files, but it was enough to validate that `gbif_download()` works great 🎉 and provide some data to work with (note I deleted the last, partial downloaded parquet file)
- I think an estimate, even rough or a minimum, re: how long it might take on "average" to download _all_ the GBIF data would be useful
- I specified where the data was stored, and then had to add the `.../occurence.parquet/` to the end of that path for `gbif_conn()`---it might be useful to add an example of this approach to the documentation:

```
gbif_download(dir = "/Users/stephhazlitt/gbif-data/")

gbif_dir <- "/Users/stephhazlitt/gbif-data/occurrence.parquet/"
conn <- gbif_conn(gbif_dir)

gbif <- tbl(conn, "gbif")
gbif
```

- Acknowledging that the intended focus of the package is around a relational database interface to parquet-based serializations of GBIF data, I think it would also be worth fleshing out the {arrow} approach to achieving the same end result---the equivalent as outlined in the remote section's arrow vs `to_duckdb = TRUE` pathways. It looks like the docs were written virtually at the same time as the {arrow} [6.0.0 release](https://arrow.apache.org/blog/2021/11/08/r-6.0.0/), which brought a lot of added {dplyr} support to {arrow} making this a reasonable alternative approach, at least with the amount of data I downloaded and the example data.
- Querying the example data using `gbif_conn()`+ `tbl()` or using `arrow::open_dataset()`:

```
library(gbifdb)
library(arrow)
library(ggplot2)

gbif_parquet <- gbif_example_data()

## database
conn_example <- gbif_conn(gbif_parquet)
gbif_example <- tbl(conn_example, "gbif")

gbif_example |>
  group_by(order, phylum) |>
  summarise(n_species = n_distinct(species)) |>
  collect() |>
  ggplot() +
  geom_col(aes(n_species, order, fill = phylum))

## arrow
open_dataset(gbif_parquet) |>
  group_by(order, phylum) |>
  summarise(n_species = n_distinct(species)) |>
  collect() |>
  ggplot() +
  geom_col(aes(n_species, order, fill = phylum))
```

- I am not sure what the current design is around the GBIF parquet files, however if at some point the GBIF parquet serializations were partitioned by common query parameters (e.g. year, month) I think the {gbifdb} remote and local queries would be (even) faster---it might be interesting to experiment with partitions.

#### Minor Comments

- It would be great to see Contribution Guidelines added as either a stand alone `.md` file or in the README.md file
- I think adding Citation information would also be beneficial, again in the README.md or a CITATION file
- My local Check Build output flagged a typo in gbif_download.R (occurence -> occurrence)
- Adding top-level package documentation so `?gbigdb` works for users is always nice in my experience
- Adding licence and code coverage badges for easy scanning would be nice, maybe when all the rOpenSci badges are added
- Including the url to {rgbif} would be a convenience for users exploring workflow options https://docs.ropensci.org/rgbif/
