### 2022-01-12 #492 @cboettig
Thanks @stephhazlitt this is super :tada: !  One design question you bring up which I'd really like advice on:

I completely hear you that it's awkward that `gbif_conn()` returns a remote connection to a database, while `gbif_remote()` returns a remote table; i.e. that in the former we need to remember to use `dplyr::tbl()` but with the arrow-based interfaces we don't.  This kind of begs for a `gbif_local()`, which would just be `tbl(gbif_conn(), "gbif")`.  So **Q1** Is it worth making `dplyr` a hard dependency in order to add `gbif_local()`?

Currently, `dplyr` is only a soft dependency, since presumably some applications or users will just want the DBI connection object for which they are happy querying in SQL (some users/apps will know SQL well and may not like using dplyr's doing it for them).  I think of `dplyr` as a heavy dependency, but it's also common so maybe I should just bite that bullet?

**Q2** I'm also obviously on the fence as to whether `gbif_remote()` should just go ahead and call `to_duckdb()` or not.  Doing so would make the interface a bit more consistent between the two connection types.  OTOH, `to_duckdb()` can actually slow things down a bit a present when doing queries that arrow already knows how to do, which as you noted is quite a lot in `arrow` 6.0.1. Thoughts on either of these?

I totally hear you on the timing and network issues. I'll add some comments in the README and function docs.  One interesting case to mention might be increasingly-popular cloud-hosted computing services, like https://rstudio.cloud.  Similarly, I teach my courses using  RStudio instances Berkeley provides that are similarly dinky in RAM and CPU, but because they are hosted on Google Cloud have access to good bandwidth.  I think when using AWS-EC2 instances in the matching region as GBIF uses would see particularly impressive performance, comparable to the local download, but almost all cloud providers have rather fast network performance, even on instances that have less RAM or CPU power than a dinky laptop.

Benchmarking network issues is tricky though, or at least I've never been properly trained how to do it!  Maybe some folks from the arrow team could advice me on that.  With colleagues over at GBIF, we put together this draft blog post very recently: https://github.com/gbif/data-blog/blob/master/content/post/2022-01-04-make-big-queries-using-GBIF-parquet-snapshots.md.  My local laptop is a crappy Window-Go tablet that struggles to run R, but I log in remotely to an rstudio session on my desktop at work where we actually have gigabit ethernet, so I see runtimes that are almost an order of magnitude faster than John reports for remote connections, and fewer connection issues.

Also good point about the laptop falling asleep on the whole-data download - can add notes to the docs about that too.  It's about a 100 GB download.  Google tells me the average download speed in the US is over 100MB/s but that sounds ridiculous, let's say you get 10 MB/s downloads it ought to take 2.7 hours I guess?

I may also be worth noting that while 100 GB isn't all that large for a laptop, it's much bigger than most typical cloud providers will provide, so `neon_remote()` may be the only option there too.
