### 2022-01-12 #492 @stephhazlitt
Great questions.

I am biased towards a fully R workflow, so not sure my thoughts are representative of your user community, however I would go with the {dplyr} dependency and aim for a symmetrical 2-function per GBIF data location design:

`gbif_remote()` - as is, arrow table connection
`gbif_remore_db()` - wrapper for above with `to_duckdb = TRUE`

`gbif_local()` - wrapper for arrow::open_dataset() to read locally stored parquet data
`gbif_local_db() - for `tbl(gbif_conn(), "gbif")`

My thinking is that this brings the maximum flexibility for users and given tools and dependencies are all changing and improving so quickly.

If you _really_ wanted to provide a lightweight version for non-{dplyr} users, perhaps you could make a one function package for `gbif_conn()` and then use that package as a dependency in {gbifdb}? I am channeling learnings from the {bcdata} experience, where for sure if we had a redo it would be more than one package.
