### 2022-01-12 #492 @cboettig
Nice, I like it!  Just hashing out function names a bit more here:

I think going from 2 to functions (`gbif_remote()` an `gbif_local()`) to four seems a bit overkill.  I imagine the end user doesn't care so much about whether it's duckdb vs arrow at the backend (particularly when functions like `to_duckdb()` make an arrow connection use duckdb or vice versa.  In particular, I'm not sure what we achieve by having the the two different flavors of `gbif_local()` / `gbif_local_db()`.  Sure the duckdb-based one supports nearly the full dbplyr functionality, while arrow doesn't do windowed operations (at least not yet), but that seems a more minor difference.  (technically, `duckdb` can access remote parquet functions directly without `arrow`, which means we could drop the `arrow` dependency and  `gbif_remote()` would match `gbif_local()` use for all cases without needing `to_duckdb()` to use windowed operations.... but so far, it's not enabled by default in the R client, https://github.com/duckdb/duckdb/issues/2177)

 re audience, I'm with you on a fully R workflow, I just meant that there's probably a subset of old-school R users out there who have been using the `DBI` R package since before dplyr/dbplyr and are probably happy to keep doing so.  But I can keep gbif_conn() as a function for them but still make `dplyr` a hard dependency I think.  (I'm not sure, but this might mean `dbplyr` should be a dependency too?)  The two-package idea is a good one, but I think overkill here since this this is already a thin wrapper around `duckdb` and `arrow`.
