### 2022-01-13 cboettig/gbifdb#2 @cboettig
@Pakillo Thanks so much for the detailed report!  I can reproduce this crash.   Looks like this might be a `duckdb` issue, possibly stemming from the nature of the parquet data with how filters work when all columns are returned.  I need to play around a bit more and probably follow up with the duckdb devs.

For comparison, can you try the same query in pure `arrow`, e.g.

```r
gbif <- arrow::open_dataset("/home/shared-data/gbif/occurrence/2021-11-01/occurrence.parquet/")
pinsapo <- gbif %>%
  filter(species == "Abies pinsapo",
         countrycode == "ES",
         year > 2010)
pinsapo %>% collect()
```

That query works for me (though is still somewhat intensive).

Also would be great if you could test with adding a `select()` call to get only a subset of columns back (both in `arrow` and the native duckdb call).  It looks like it makes a huge difference to select() only the columns you need when querying parquet (for both local duckdb and remote arrow connections).

(Obviously these are things we'd want to at least document!  and it's possible we can avoid them, at least on the local side, by reading the parquet files into the native duckdb backend first.  I'm totally still learning here so thanks for exploring this with me!)
