### 2022-01-13 cboettig/gbifdb#2 @Pakillo
Thanks for the quick reply @cboettig!

You're totally right about `select`. Once I select just a few columns to return, it works! And quite fast.

```r
con <- gbif_conn(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/")
gbif <- tbl(con, "gbif")

pinsapo <- gbif %>%
  filter(species == "Abies pinsapo",
         countrycode == "ES",
         year > 2010) %>%
  select(species, year, decimallatitude, decimallongitude)

pinsapo
```

BUT from several tests I've done selecting different columns, it looks to me the problem may be with the last two columns in the database: `mediatype` and `issue`. I see [here](https://github.com/gbif/occurrence/blob/dev/aws-public-data.md) that these two columns are special (string arrays), compared to all the other columns, that are strings, double, etc.

So, I can select up to 48 columns from the 50 available without any problem, as long as I don't select any of those two columns.

Does this make sense? What can be done? (apart from documenting this in the package, of course). If you avoid those two columns it seems all my queries work fine
