### 2022-01-13 36dea0d
Add NEWS.md
