+++ R/gbif_conn.R
@@ -13,10 +13,28 @@
 #' con <- gbif_conn(gbif.parquet)
 #' 
 gbif_conn <- function(dir = gbif_dir(), tblname = "gbif") {
-  
-  parquet <- file.path(dir, "*")
+  gbif_conn_arrow(dir, tblname)
+}
+
+# Using arrow as the backend for local data gives better performance
+gbif_conn_arrow <- function(dir = gbif_dir(), 
+                             tblname = "gbif") {
+  arrow <- arrow::open_dataset(dir)
   conn <- DBI::dbConnect(duckdb::duckdb())
-  if(! tblname %in% DBI::dbListTables(conn)){
+  duckdb::duckdb_register_arrow(conn, tblname, arrow)
+  conn
+}
+
+
+
+
+
+## Pure duckdb -- currently crashes if all columns are selected
+gbif_conn_duckdb <- function(dir = gbif_dir(), 
+                             tblname = "gbif") {
+  parquet <- file.path(dir, "*")
+  conn <- DBI::dbConnect(duckdb::duckdb(), database)
+  if (! tblname %in% DBI::dbListTables(conn)){
     query <- paste0("CREATE VIEW '", tblname,
                   "' AS SELECT * FROM parquet_scan('",
                   parquet, "');")
@@ -25,4 +43,4 @@ gbif_conn <- function(dir = gbif_dir(), tblname = "gbif") {
   conn
 }
 
-
+## CREATE TABLE with persistent storage currently crashes!
