+++ NEWS.md
@@ -1,0 +1,3 @@
+# gbifdb 0.1.1
+
+* Added a `NEWS.md` file to track changes to the package.
