+++ DESCRIPTION
@@ -25,10 +25,10 @@ Depends: R (>= 4.0)
 Imports:
     arrow (>= 6.0.1),
     duckdb (>= 0.2.9),
-    DBI
+    DBI,
+    dplyr
 Suggests:
     spelling,
-    dplyr,
     dbplyr,
     testthat (>= 3.0.0),
     covr,
