+++ NAMESPACE
@@ -4,4 +4,6 @@ export(gbif_conn)
 export(gbif_dir)
 export(gbif_download)
 export(gbif_example_data)
+export(gbif_local)
 export(gbif_remote)
+importFrom(dplyr,tbl)
