### 2022-01-13 cboettig/gbifdb#2 @cboettig
nice tracking that down!  yeah that makes sense!  Yeah, I think we can potentially work around this issue then by excluding those 2 columns by default (e.g. internally inside `gbif_remote()` and `gbif_local()` ?  (Thanks to @stephhazlitt I've added a `gbif_local()` as a more intuitive parallel to `gbif_remote()`; lives on the `patch-1` branch for now.  Also `gbif_local()` is now using duckdb-via-arrow as the default backend instead of pure duckb since it seems a bit more robust to this issue)
