### 2022-01-13 #492 @Pakillo
## {gbifdb} Package Review

---

Thanks a lot for the invitation to review this package. I hesitated to accept the invitation at first. First, I doubt I can ever improve anything @cboettig has written. Second, although I have some modest experience using databases with `dplyr`, I have never used `arrow` or `duckdb`. But, I have learnt a lot reviewing `gbifdb` and, as a frequent GBIF user, my own work will benefit directly from this exercise (and the package itself, of course). I do hope my review is useful, at least from the perspective of an intermediate R user keen to start using the package.

The README/vignette is already very informative and nice. Some suggestions for improving it:

- I think it would be good to mention somewhere that this GBIF dataset is a snapshot of "all CC0 and CC-BY licensed data in GBIF that have coordinates which passed automated quality checks", as explained [here](https://github.com/gbif/occurrence/blob/master/aws-public-data.md). From there I understand that occurrence records without coordinates will not be included in such dataset. I think some future users may be tricked by this, e.g. if they query number of records for some species over time, ignoring that records without coordinates will be missing.

- "`gbifdb` has few dependencies: only duckdb and DBI are required". But the DESCRIPTION file also imports `arrow`?

- "gbifdb currently requires the dev version of duckdb". Is this still correct? I installed the latest `duckdb` version from GitHub as suggested (version 0.3.2). Yet the DESCRIPTION file specifies duckdb (>= 0.2.9). And CRAN already has version 0.3.1-1. Installing duckdb from github took quite a few minutes in my Ubuntu laptop, so if the CRAN version is already good, it would nice to update the README instructions accordingly. Otherwise update the version requirement in DESCRIPTION Imports?

- From our tests and also [your post at the GBIF data blog](https://github.com/gbif/data-blog/blob/cf9be1fc37f809c599a9dd6623324aa67f6712f5/content/post/2022-01-04-make-big-queries-using-GBIF-parquet-snapshots.md), it looks like `gbifdb` can be much much slower than `rgbif` for certain queries, e.g. retrieving occurrence records for a single species. I did not expect such a striking difference at first. I think it would be good to warn users about this already in the README. That is, for some use cases (like retrieving records for a single/few species or countries), it may be better to use `rgbif` than `gbifdb`, right?

- From my tests, it also looks like choosing the closest AWS region can speed up queries (and downloading) considerably. Hence, it might be nice to mention this already in the README, e.g. when first showing the use of `gbif_remote`, including an example with a different region than the default (also, is "af-south-1" the best default?). Otherwise many people may overlook this and wait unnecessarily long times (happened to me at first). The help of `gbif_remote` could also be more informative about how to choose the best region in each case (eg. see https://github.com/Pakillo/gbifdb/commit/e34d465fd97e45515f61d33ff9f90dda5a2f001f).

- When showing the use of `gbif_remote`, it might be useful to include an entire example using `collect`, for those users not so familiar with databases, e.g.

```
gbif %>%
  filter(species == "Abies pinsapo",
         countrycode == "ES",
         year > 2000) %>%
  count(year) %>%
  collect()

# A tibble: 13 × 2
    year     n
   <int> <int>
 1  2017     4
 2  2020    44
 3  2018    24
 4  2019    61
 5  2016     6
 6  2021     7
 7  2006     5
 8  2004     1
 9  2015     3
10  2014     2
11  2013     2
12  2009     1
13  2010     1
```

- Also, I think it would be nice to include a link to [this table](https://github.com/gbif/occurrence/blob/master/aws-public-data.md) describing all the available columns. Otherwise it can be hard to figure out how to filter a dataset, which variables can be used, etc.

#### Downloading the database

`gbif_download` looks really handy, but I struggled a bit to download the entire database, even using a good (fiber) internet connection.

This is the code I used to download:

```r
gbif_download(dir = "/media/frs/Elements/gbifdb/",
  region = "eu-central-1",
  bucket = "gbif-open-data-eu-central-1")
```

My first attempt failed within 20 minutes:

```r
# <== Saving object 'occurrence/2021-11-01/occurrence.parquet/000129' to '/media/frs/Elements/gbifdb///occurrence.parquet/000129'
# Error in curl::curl_fetch_disk(url, x$path, handle = handle) :
#   Timeout was reached: [s3-eu-central-1.amazonaws.com] Resolving timed out after 10000 milliseconds
```

A second attempt gave an error after a while:

```r
# <== Saving object 'occurrence/2021-11-01/occurrence.parquet/001044' to '/media/frs/Elements/gbifdb///occurrence.parquet/001044'
# 131 local files not found in bucket 'gbif-open-data-eu-central-1'
# List of 4
# $ Code     : chr "AccessDenied"
# $ Message  : chr "Access Denied"
# $ RequestId: chr "CA2X7EK3FCXH7RP2"
# $ HostId   : chr "9h1TicaLFACsutywE7S7etD/yuYckCyxHKbnDfL7zd7hJv2PTCqPxkV8m7gwtxKCtWPCnQEPca0="
# - attr(*, "headers")=List of 7
# ..$ x-amz-request-id : chr "CA2X7EK3FCXH7RP2"
# ..$ x-amz-id-2       : chr "9h1TicaLFACsutywE7S7etD/yuYckCyxHKbnDfL7zd7hJv2PTCqPxkV8m7gwtxKCtWPCnQEPca0="
# ..$ content-type     : chr "application/xml"
# ..$ transfer-encoding: chr "chunked"
# ..$ date             : chr "Thu, 13 Jan 2022 11:22:45 GMT"
# ..$ server           : chr "AmazonS3"
# ..$ connection       : chr "close"
# ..- attr(*, "class")= chr [1:2] "insensitive" "list"
# - attr(*, "class")= chr "aws_error"
# NULL
# Error in parse_aws_s3_response(r, Sig, verbose = verbose) :
#   Forbidden (HTTP 403).
```

Third attempt also gave similar error, even though it looks like all the files had been successfully downloaded (1045 parquet files plus the citation file):

```r
# <== Saving object 'occurrence/2021-11-01/occurrence.parquet/001044' to '/media/frs/Elements/gbifdb///occurrence.parquet/001044'
# 1046 local files not found in bucket 'gbif-open-data-eu-central-1'
# List of 4
#  $ Code     : chr "AccessDenied"
#  $ Message  : chr "Access Denied"
#  $ RequestId: chr "21VETY51D6FQ6K3A"
#  $ HostId   : chr "couGo2j17BhB5a8k3GFjI8gIrO3akAbRzCaikOOk30sEIr7d4nCwBLcZRdA0wlYAwGF969q2KZ4="
#  - attr(*, "headers")=List of 7
#   ..$ x-amz-request-id : chr "21VETY51D6FQ6K3A"
#   ..$ x-amz-id-2       : chr "couGo2j17BhB5a8k3GFjI8gIrO3akAbRzCaikOOk30sEIr7d4nCwBLcZRdA0wlYAwGF969q2KZ4="
#   ..$ content-type     : chr "application/xml"
#   ..$ transfer-encoding: chr "chunked"
#   ..$ date             : chr "Thu, 13 Jan 2022 13:55:31 GMT"
#   ..$ server           : chr "AmazonS3"
#   ..$ connection       : chr "close"
#   ..- attr(*, "class")= chr [1:2] "insensitive" "list"
#  - attr(*, "class")= chr "aws_error"
# NULL
# Error in parse_aws_s3_response(r, Sig, verbose = verbose) :
#   Forbidden (HTTP 403).
```

In general, asking from ignorance, is there any way to make the download easier/less fragile to the user? Also, it seems the download starts from scratch every time you call `gbif_download` (that is, files already on disk are replaced again, even if requesting the same GBIF snapshot)?

- Also perhaps `gbif_download` documentation may be improved a bit, e.g. see https://github.com/Pakillo/gbifdb/commit/0ddc505b4626e158871643fafe991a795d3c77a6.

#### Querying the local database

- As @stephhazlitt noted, it would be nice to add an example to `gbif_conn` with a user-provided dir. My first attempt failed:

```r
con <- gbif_conn(dir = "/media/frs/Elements/gbifdb/")
Error in .local(conn, statement, ...) :
  duckdb_prepare_R: Failed to prepare query CREATE VIEW 'gbif' AS SELECT * FROM parquet_scan('/media/frs/Elements/gbifdb//*');
Error: Invalid Input Error: No magic bytes found at end of file '/media/frs/Elements/gbifdb/citation.txt'
```

You have to add "occurrence.parquet" for it to work:

```r
con <- gbif_conn(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/")
```

- Once connected to the database, I was able to reproduce all the examples in the readme/vignette. I also managed to query some data for a random species:

``` r
library(gbifdb)

con <- gbif_conn(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/")
gbif <- tbl(con, "gbif")

pinsapo <- gbif %>%
  filter(species == "Abies pinsapo",
         countrycode == "ES",
         year > 2010) %>%
  select(species, year, decimallatitude, decimallongitude)

pinsapo
#> # Source:   lazy query [?? x 4]
#> # Database: duckdb_connection
#>    species        year decimallatitude decimallongitude
#>    <chr>         <int>           <dbl>            <dbl>
#>  1 Abies pinsapo  2013            36.7            -4.96
#>  2 Abies pinsapo  2013            36.8            -5.39
#>  3 Abies pinsapo  2021            36.7            -5.04
#>  4 Abies pinsapo  2018            40.5            -4.1
#>  5 Abies pinsapo  2020            40.9            -4
#>  6 Abies pinsapo  2019            41.7            -3.7
#>  7 Abies pinsapo  2021            36.7            -5.04
#>  8 Abies pinsapo  2021            36.7            -5.04
#>  9 Abies pinsapo  2019            40.5            -3.3
#> 10 Abies pinsapo  2019            40.7            -4.7
#> # … with more rows
```

However, there seems to be a problem when two special columns (`mediatype` and `issue`) are selected for return in the results table, causing R session to crash (see https://github.com/cboettig/gbifdb/issues/2). I hope this will be easy to fix. Otherwise it should be warned in the package documentation.

**Minor comments**

- The pkgdown website (https://cboettig.github.io/gbifdb) is not up to date with the latest changes in the package.
