### 2022-01-13 cboettig/gbifdb#2 @Pakillo
Excluding those two columns looks like a good solution to me, at least by now :+1:

I had submitted the review before seeing your response, just to move things forward, but I'm happy to keep an eye if I can be of help

Cheers
