+++ tests/testthat/test_gbifdb.R
@@ -7,6 +7,17 @@ test_that("gbif_example_data()", {
 
 })
 
+test_that("gbif_conn() duckdb", {
+  
+  skip_on_os("solaris")
+  path <- gbif_example_data()
+  conn <- gbif_conn(path, backend="duckdb")
+  
+  expect_true(dir.exists(path))
+  expect_true(inherits(conn, "duckdb_connection"))
+  
+})
+
 test_that("gbif_conn()", {
   
   skip_on_os("solaris")
@@ -16,6 +27,18 @@ test_that("gbif_conn()", {
   expect_true(dir.exists(path))
   expect_true(inherits(conn, "duckdb_connection"))
 
+})
+
+test_that("gbif_local()", {
+  
+  skip_on_os("solaris")
+  path <- gbif_example_data()
+  gbif <- gbif_local(path)
+  
+  expect_true(inherits(gbif, "tbl"))
+  expect_true(inherits(gbif, "tbl_dbi"))
+  expect_true(inherits(gbif, "tbl_duckdb_connection"))
+  
 })
 
 test_that("gbif_dir()", {
@@ -33,7 +56,8 @@ test_that("gbif_remote()", {
   skip_if_not(has_s3)
 
   conn <- gbif_remote(to_duckdb = FALSE)
-  expect_true(inherits(conn, "Dataset"))
+  expect_true(inherits(conn, "arrow_dplyr_query"))
+  #expect_true(inherits(conn, "Dataset"))
 })
 
 test_that("gbif_remote(to_duckdb=TRUE)....slow!", {
