### 2022-01-14 48f3d60
export gbif_version() fix detection of local path
