+++ R/gbif_conn.R
@@ -13,7 +13,8 @@
 #' gbif.parquet <- gbif_example_data()
 #' con <- gbif_conn(gbif.parquet)
 #' 
-gbif_conn <- function(dir = gbif_dir(),
+gbif_conn <- function(dir = gbif_parquet_dir(version = 
+                                               gbif_version(local=TRUE)),
                       tblname = "gbif", 
                       backend = c("arrow", "duckdb")) {
   backend <- match.arg(backend)
@@ -24,7 +25,7 @@ gbif_conn <- function(dir = gbif_dir(),
 }
 
 # Using arrow as the backend for local data gives better performance
-gbif_conn_arrow <- function(dir = gbif_dir(), 
+gbif_conn_arrow <- function(dir, 
                              tblname = "gbif") {
   arrow <- arrow::open_dataset(dir)
   conn <- cachable_duckdb_conn()
@@ -48,7 +49,7 @@ cachable_duckdb_conn <- function(){
 
 ## Pure duckdb -- currently crashes if all columns are selected
 ## Currently not used!
-gbif_conn_duckdb <- function(dir = gbif_dir(), 
+gbif_conn_duckdb <- function(dir, 
                              tblname = "gbif") {
   parquet <- file.path(dir, "*")
   conn <- cachable_duckdb_conn()
