### 2022-01-14 #492 @cboettig
@Pakillo Thanks for this amazing review (and the help tracking down the issues with in https://github.com/cboettig/gbifdb/issues/2)!  Still working through it, but wanted to jot down what I've tried to address so far.

- I've updated the default bucket to `us-east-1` instead of South Africa (A biased default to be sure, but also what I believe most AWS clients default to when nothing is set). I've tried to note in the docs that users probably want to set the bucket accordingly based on their location.

- Not entirely sure what was up with `gbif_download()`, I'll test more too, but I've made a few tweaks.  Like you say, setting a closer region helps, and I've made sure to unset alternative credentials more aggressively.  My understanding of `aws.s3::s3sync()` behavior is that it is checking md5sums so should not re-download, but I haven't tested to confirm that.  Now that `arrow` is a dependency I'm considering using it's S3-interface to do the download instead, but it is much more low-level.  It's quite possible that the `awscli` tool (as shown at https://registry.opendata.aws/gbif/) at or other such tools give significantly better performance.  Maybe I should point users at that and deprecate this function...

I think I've addressed some of the performance issues you identified in gbif remote now.  May still be slower than rgbif for small queries, (again depending on network speed) but will need to benchmark more.

more to come, gotta run
