### 2022-01-17 #492 @maelle
@jhollist this might be a small buffy bug, thanks for your patience :-) https://github.com/ropensci-org/buffy/issues/54
