### 2022-01-21 #492 @jhollist
Sorry for delay in moving things along, @maelle!  Looks like I had a bug too!  Works now.
