### 2022-01-21 #492 @cboettig
Thanks all!  I'm working on revisions still as noted above, will include a complete report & links to my PR diffs when I've ticked off everything.
