+++ README.Rmd
@@ -116,16 +116,11 @@ gbif_download()
 ```
 
 By default, this will download to the dir given by `gbif_dir()`.  
-An alternative directory can be set as an argument, or by setting the path in the environmental variable, `GBIF_HOME`.
+An alternative directory can be provided by setting the environmental variable, `GBIF_HOME`,
+or providing the path to the directory containing the parquet files directly.
 
 Once you have downloaded the parquet-formatted GBIF data, 
-simply point `gbif_local()` at the directory containing your parquet files to initialize a connection.
-
-
-
-
-
-The resulting connection can be used with `dplyr::tbl()` to access the full gbif data:
+`gbif_local()` will establish a connection to thes local parquet files. 
 
 ```{r}  
 gbif <- gbif_local()
