### 2022-02-02 #492 @cboettig
Ok, I _think_ I've hit everything now, beween https://github.com/cboettig/gbifdb/pull/3 and https://github.com/cboettig/gbifdb/pull/4 . Summaries on each PR, and maybe the diffs are helpful too.
