### 2022-02-04 #492 @jhollist
@Pakillo and @stephhazlitt just wanted to make sure you saw these.  @cboettig has two PR's linked in https://github.com/ropensci/software-review/issues/492#issuecomment-1027489887 that should address the issues raised in your reviews.  Could you please take a look at these and let us know if everything is good to go?
