### 2022-02-12 #492 @Pakillo
Hi,

Thanks @cboettig for working out all those changes. The documentation and usage look much nicer.

Unfortunately I'm still having issues trying to make the package work...

## Remote connection

First, if I try the remote connection (gbif_remote) using duckdb (the default), my R session invariable crashes whenever I try a consult, as it goes out of RAM (16 GB in my laptop).

So, I can connect successfully to the remote database:
`gbif <- gbif_remote(bucket = "gbif-open-data-eu-central-1", to_duckdb = TRUE)`

but when I try a consult, like the one in the README or others, my R session crashes:
```r
gbif %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
```

The remote connection does work well if I choose an arrow connection. Although consulting the database keeps being a bit slow?

``` r
library(gbifdb)
library(dplyr)  # optional, for dplyr-based operations

gbif <- gbif_remote(bucket = "gbif-open-data-eu-central-1",
                    to_duckdb = FALSE)

system.time({
  ex <- gbif %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
})
#>    user  system elapsed
#>   0.055   0.000   0.056

system.time({
  ex %>%
    collect()
})
#>    user  system elapsed
#> 310.606  28.141 126.655

system.time({
df <- gbif %>%
  filter(species == "Abies pinsapo",
         countrycode == "ES",
         year > 2020) %>%
  # count(year) %>%
  select(species, year, locality, decimallatitude, decimallongitude)
})
#>    user  system elapsed
#>   0.033   0.003   0.036

system.time({
df %>%
  collect()
})
#>    user  system elapsed
#> 328.496 115.479 426.916

system.time({
df <- gbif %>%
  filter(species == "Abies pinsapo",
         countrycode == "ES",
         year > 2000) %>%
  count(year) %>%
  collect()
})
#>    user  system elapsed
#> 271.148  43.273 220.665

```

<summary>
Session info
</summary>

``` r
sessioninfo::session_info()
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 4.1.2 (2021-11-01)
#>  os       Ubuntu 20.04.3 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language en_GB:en
#>  collate  en_GB.UTF-8
#>  ctype    en_GB.UTF-8
#>  tz       Europe/Madrid
#>  date     2022-02-12
#>  pandoc   2.11.4 @ /usr/lib/rstudio/bin/pandoc/ (via rmarkdown)
#>
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version date (UTC) lib source
#>  arrow         7.0.0   2022-02-10 [1] RSPM (R 4.1.0)
#>  assertthat    0.2.1   2019-03-21 [1] RSPM (R 4.1.0)
#>  bit           4.0.4   2020-08-04 [1] RSPM (R 4.1.0)
#>  bit64         4.0.5   2020-08-30 [1] RSPM (R 4.1.0)
#>  cli           3.1.1   2022-01-20 [1] RSPM (R 4.1.0)
#>  crayon        1.4.2   2021-10-29 [1] CRAN (R 4.1.1)
#>  DBI           1.1.2   2021-12-20 [1] RSPM (R 4.1.0)
#>  digest        0.6.29  2021-12-01 [1] RSPM (R 4.1.0)
#>  dplyr       * 1.0.8   2022-02-08 [1] RSPM (R 4.1.0)
#>  ellipsis      0.3.2   2021-04-29 [1] RSPM (R 4.1.0)
#>  evaluate      0.14    2019-05-28 [1] RSPM (R 4.1.0)
#>  fansi         1.0.2   2022-01-14 [1] RSPM (R 4.1.0)
#>  fastmap       1.1.0   2021-01-25 [1] RSPM (R 4.1.0)
#>  fs            1.5.2   2021-12-08 [1] RSPM (R 4.1.0)
#>  gbifdb      * 0.1.1   2022-02-12 [1] Github (cboettig/gbifdb@7f425f5)
#>  generics      0.1.2   2022-01-31 [1] RSPM (R 4.1.0)
#>  glue          1.6.1   2022-01-22 [1] CRAN (R 4.1.2)
#>  highr         0.9     2021-04-16 [1] RSPM (R 4.1.0)
#>  htmltools     0.5.2   2021-08-25 [1] RSPM (R 4.1.0)
#>  knitr         1.37    2021-12-16 [1] RSPM (R 4.1.0)
#>  lifecycle     1.0.1   2021-09-24 [1] RSPM (R 4.1.0)
#>  magrittr      2.0.2   2022-01-26 [1] RSPM (R 4.1.0)
#>  pillar        1.7.0   2022-02-01 [1] RSPM (R 4.1.0)
#>  pkgconfig     2.0.3   2019-09-22 [1] RSPM (R 4.1.0)
#>  purrr         0.3.4   2020-04-17 [1] RSPM (R 4.1.0)
#>  R6            2.5.1   2021-08-19 [1] RSPM (R 4.1.0)
#>  reprex        2.0.1   2021-08-05 [1] RSPM (R 4.1.0)
#>  rlang         1.0.1   2022-02-03 [1] RSPM (R 4.1.0)
#>  rmarkdown     2.11    2021-09-14 [1] RSPM (R 4.1.0)
#>  rstudioapi    0.13    2020-11-12 [1] RSPM (R 4.1.0)
#>  sessioninfo   1.2.2   2021-12-06 [1] RSPM (R 4.1.0)
#>  stringi       1.7.6   2021-11-29 [1] RSPM (R 4.1.0)
#>  stringr       1.4.0   2019-02-10 [1] RSPM (R 4.1.0)
#>  tibble        3.1.6   2021-11-07 [1] RSPM (R 4.1.0)
#>  tidyselect    1.1.1   2021-04-30 [1] RSPM (R 4.1.0)
#>  utf8          1.2.2   2021-07-24 [1] RSPM (R 4.1.0)
#>  vctrs         0.3.8   2021-04-29 [1] RSPM (R 4.1.0)
#>  withr         2.4.3   2021-11-30 [1] RSPM (R 4.1.0)
#>  xfun          0.29    2021-12-14 [1] RSPM (R 4.1.0)
#>  yaml          2.2.2   2022-01-25 [1] CRAN (R 4.1.2)
#>
#>  [1] /home/frs/R/x86_64-pc-linux-gnu-library/4.0
#>  [2] /opt/R/4.1.2/lib/R/library
#>
#> ──────────────────────────────────────────────────────────────────────────────
```

## Local database

Querying the local database does work well, and quite fast, but only when using the `duckdb` backend:

``` r
library(gbifdb)
library(dplyr)

gbif <- gbif_local(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/",
                   backend = "duckdb")

system.time({
  ex <- gbif %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
  ex
})
#>    user  system elapsed
#>   0.015   0.000   0.015
```

If I choose the `arrow` backend, my R session always crashes (again due to RAM)

Sorry for the bad news. Please let me know if I'm doing something wrong. Otherwise I hope it's not too complicated to fix.

Cheers

P.S. The pkgdown website and the README on github are out of date as `README.Rmd` has not been knitted.
