### 2022-02-14 #492 @cboettig
Thanks @Pakillo !  Yup, looks like there are still upstream issues in `arrow` and `duckdb` being worked out -- e.g. `arrow` 7.0 R client has a bug in `count` that causes it to read into memory.   I have a handful of open tickets in `arrow` and they have a great and very responsive team of developers, so I'm hopeful this will improve soon.  See:
- https://issues.apache.org/jira/browse/ARROW-14745
- https://issues.apache.org/jira/browse/ARROW-15081
- https://issues.apache.org/jira/browse/ARROW-13761

bizarrely, I don't see `duckdb` version in your `sessionInfo`.  I have an outstanding issue on `duckdb` regarding RAM performance on `to_duckdb()` as well https://github.com/duckdb/duckdb/issues/2768, so I think we may need to wait for that one to resolve upstream as well.

It sounds like we have reasonable functionality as long as a user sticks with `arrow` for remote connections and `duckdb` for local connections then?  (Ironically I've chosen the opposite setup as the default both times!)  I'm still optimistic things will improve with `arrow` and `duckdb` as the upstream issues are closed.

- [ ] I'll update pkgdown and README.md
- [ ] I'll add the package-level doc-string (thanks @stephhazlitt , I always forget to do that one)
