### 2022-02-15 #492 @Pakillo
Thanks @cboettig. I reckon these issues stem from upstream dependencies (arrow and duckdb) - hope they are fixed soon.

> bizarrely, I don't see duckdb version in your sessionInfo

That's because that in that session I am choosing directly an arrow backend (`to_duckdb = FALSE`), so it seems duckdb does not get loaded.

@stephhazlitt Thanks for the suggestion! Using `arrow::open_dataset`, as you suggest, does not crash my R session (while using gbif_local with `backend = "arrow"` does crash it). Why could that be @cboettig?

Still, arrow seems to be slower than duckdb when querying my local database:

``` r
system.time({
  ex <- arrow::open_dataset("/media/frs/Elements/gbifdb/occurrence.parquet/") %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
  ex
})
#>    user  system elapsed
#>   0.210   0.021   0.269
```

``` r
gbif <- gbif_local(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/",
                   backend = "duckdb")

system.time({
  ex <- gbif %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
  ex
})
#>    user  system elapsed
#>   0.014   0.000   0.014
```
