### 2022-02-15 #492 @cboettig
> Why could that be @cboettig?

`gbif_local()` using the arrow backend calls `arrow::open_dataset()`, but then takes the additional step of registering the connection with `duckdb`:

https://github.com/cboettig/gbifdb/blob/1b50eb6d83963d2c10e3cfb40ef2ff8fd2df57c3/R/gbif_conn.R#L28-L32

This ought to have an effect similar to `arrow::to_duckdb()` , but is done via the `duckdb::duckdb_register_arrow` instead.  This more closely parallels the user experience of `backend="duckdb"`.  In particular, this means a user can take advantage of all methods, including window functions, that are supported in `duckdb`, and also makes the argument compatible with `gbif_conn(backend="duckdb")`, that is, you can get back a connection object. That could be useful if you wanted to database-like operations instead of single-table operations (e.g. join this table on another table in the same connection).

Might be good for us to report the crash upstream to `duckdb`.  What version of `duckdb` do you get these crashes on?
