### 2022-02-15 #492 @Pakillo
I see... Thanks for the explanation! Then the issues seem to be with duckdb rather than arrow, right?

I'm using latest `duckdb` available on CRAN (0.3.2)

This is the code that crashes my session:

```r
library(gbifdb)
library(dplyr)

gbif <- gbif_local(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/",
                   backend = "arrow")

# This bit below crashes R:
gbif %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
```

Session info:

```r
─ Session info ───────────────────────────────────────────────────────────────────────────────────────────────
 setting  value
 version  R version 4.1.2 (2021-11-01)
 os       Ubuntu 20.04.3 LTS
 system   x86_64, linux-gnu
 ui       RStudio
 language en_GB:en
 collate  en_GB.UTF-8
 ctype    en_GB.UTF-8
 tz       Europe/Madrid
 date     2022-02-15
 rstudio  1.4.1717 Juliet Rose (desktop)
 pandoc   2.5 @ /usr/bin/pandoc

─ Packages ───────────────────────────────────────────────────────────────────────────────────────────────────
 package       * version date (UTC) lib source
 arrow           7.0.0   2022-02-10 [1] RSPM (R 4.1.0)
 assertthat      0.2.1   2019-03-21 [1] RSPM (R 4.1.0)
 aws.s3          0.3.21  2020-04-07 [1] RSPM (R 4.1.0)
 aws.signature   0.6.0   2020-06-01 [1] RSPM (R 4.1.0)
 base64enc       0.1-3   2015-07-28 [1] RSPM (R 4.1.0)
 bit             4.0.4   2020-08-04 [1] RSPM (R 4.1.0)
 bit64           4.0.5   2020-08-30 [1] RSPM (R 4.1.0)
 cli             3.1.1   2022-01-20 [1] RSPM (R 4.1.0)
 codemeta        0.1.1   2021-12-22 [1] RSPM (R 4.1.0)
 crayon          1.4.2   2021-10-29 [1] CRAN (R 4.1.1)
 curl            4.3.2   2021-06-23 [1] RSPM (R 4.1.0)
 DBI             1.1.2   2021-12-20 [1] RSPM (R 4.1.0)
 dbplyr          2.1.1   2021-04-06 [1] RSPM (R 4.1.0)
 digest          0.6.29  2021-12-01 [1] RSPM (R 4.1.0)
 dplyr         * 1.0.8   2022-02-08 [1] RSPM (R 4.1.0)
 duckdb          0.3.2   2022-02-07 [1] RSPM (R 4.1.0)
 ellipsis        0.3.2   2021-04-29 [1] RSPM (R 4.1.0)
 fansi           1.0.2   2022-01-14 [1] RSPM (R 4.1.0)
 gbifdb        * 0.1.1   2022-02-12 [1] Github (cboettig/gbifdb@7f425f5)
 generics        0.1.2   2022-01-31 [1] RSPM (R 4.1.0)
 glue            1.6.1   2022-01-22 [1] CRAN (R 4.1.2)
 httr            1.4.2   2020-07-20 [1] RSPM (R 4.1.0)
 jsonlite        1.7.3   2022-01-17 [1] RSPM (R 4.1.0)
 lifecycle       1.0.1   2021-09-24 [1] RSPM (R 4.1.0)
 magrittr        2.0.2   2022-01-26 [1] RSPM (R 4.1.0)
 pillar          1.7.0   2022-02-01 [1] RSPM (R 4.1.0)
 pkgconfig       2.0.3   2019-09-22 [1] RSPM (R 4.1.0)
 purrr           0.3.4   2020-04-17 [1] RSPM (R 4.1.0)
 R6              2.5.1   2021-08-19 [1] RSPM (R 4.1.0)
 rlang           1.0.1   2022-02-03 [1] RSPM (R 4.1.0)
 sessioninfo     1.2.2   2021-12-06 [1] RSPM (R 4.1.0)
 tibble          3.1.6   2021-11-07 [1] RSPM (R 4.1.0)
 tidyselect      1.1.1   2021-04-30 [1] RSPM (R 4.1.0)
 utf8            1.2.2   2021-07-24 [1] RSPM (R 4.1.0)
 vctrs           0.3.8   2021-04-29 [1] RSPM (R 4.1.0)
 xml2            1.3.3   2021-11-30 [1] RSPM (R 4.1.0)

 [1] /home/frs/R/x86_64-pc-linux-gnu-library/4.0
 [2] /opt/R/4.1.2/lib/R/library
```
