### 2022-02-15 #492 @cboettig
Thanks!  Yes, at the moment, it does look like purely `arrow`-based methods are working more-or-less as expected, e.g. `gbif_remote()` with `to_duckdb()` = FALSE, or a local connection using `arrow` without using `duckdb::duckdb_register_arrow`.

I think both of these _duckdb_ issues may be related to the still-open bug in the _arrow_ package, https://issues.apache.org/jira/browse/ARROW-14745 , but honestly not sure, that's above my pay grade!  I will try and follow up with the duckdb team.

I do think there's a few outstanding issues in both `arrow` and `duckdb` that I'd like to see resolved to feel better about either of these approaches.

On the test:

```r
system.time({
  ex <- arrow::open_dataset("/media/frs/Elements/gbifdb/occurrence.parquet/") %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
  ex
})
```
I think the implicit `print` command at the end should force evaluation, but I'm curious if you could just add an explicit call  to `collect()` (or `compute()`) to confirm it works as expected without crashing?

Also, I'm curious if you call `arrow::to_duckdb()` on `ex`, does it work or crash?

For now it seems like it might be reasonable for `gbif_local(backend="arrow")` to  avoid calling `duckdb::duckdb_register_arrow` at all.  This isn't quite as tidy a solution, since it means that `gbif_local()` won't be just a thin wrapper around `gbif_conn()`, but that should be invisible to the user anyway.  I can also set `to_duckdb=FALSE` once again as the default in `gbif_remote()`, thus both functions will give a "pure arrow" experience.  (I'll also have to add something to deal with the less-desirable print method of `open_dataset()` which does not return a `tbl` class object, and can make `duckdb` dependency optional then since it's not touched by default settings.

I will try and do some more benchmarking as well.
