### 2022-02-16 #492 @Pakillo
Thanks @cboettig

> I'm curious if you could just add an explicit call to collect()

I can confirm using `arrow::open_dataset` works well, also using `collect`:

``` r
library(gbifdb)
library(dplyr)

arrow::open_dataset("/media/frs/Elements/gbifdb/occurrence.parquet/") %>%
  filter(phylum == "Chordata", year > 2010) %>%
  count(class, year) %>%
  collect()
#> # A tibble: 163 × 3
#> # Groups:   class [15]
#>    class           year        n
#>    <chr>          <int>    <int>
#>  1 Aves            2015 71131647
#>  2 Reptilia        2012    57146
#>  3 Reptilia        2015    58420
#>  4 Reptilia        2013    91989
#>  5 Reptilia        2017    64181
#>  6 Mammalia        2018   450071
#>  7 Mammalia        2013   492888
#>  8 Actinopterygii  2012  1487182
#>  9 Reptilia        2011    48347
#> 10 Mammalia        2019   452761
#> # … with 153 more rows
```

> Also, I'm curious if you call arrow::to_duckdb() on ex, does it work or crash?

It works!

``` r
library(gbifdb)
library(dplyr)

ex <- arrow::open_dataset("/media/frs/Elements/gbifdb/occurrence.parquet/") %>%
  filter(phylum == "Chordata", year > 2010) %>%
  count(class, year)

ex.duckdb <- arrow::to_duckdb(ex)
ex.duckdb
#> # Source:   table<arrow_001> [?? x 3]
#> # Database: duckdb_connection
#> # Groups:   class
#>    class           year        n
#>    <chr>          <int>    <dbl>
#>  1 Aves            2015 71131647
#>  2 Reptilia        2012    57146
#>  3 Reptilia        2015    58420
#>  4 Reptilia        2013    91989
#>  5 Reptilia        2017    64181
#>  6 Mammalia        2018   450071
#>  7 Mammalia        2013   492888
#>  8 Actinopterygii  2012  1487182
#>  9 Reptilia        2011    48347
#> 10 Mammalia        2019   452761
#> # … with more rows
```
