### 2022-02-16 #492 @jonkeane
> I think both of these duckdb issues may be related to the still-open bug in the arrow package, https://issues.apache.org/jira/browse/ARROW-14745 , but honestly not sure, that's above my pay grade! I will try and follow up with the duckdb team.

That issue is closed (though the fix for it is not in the 7.0 release!) It's possible that the crash is being caused by the behavior that's in 7.0 still. Though I will note that that change is only for going from duckdb into arrow (e.g. via `to_arrow()`).

I haven't looked at the code in GBIF in detail, but at first glance I did not see code that does that. Am I missing something? Or is it that after being read in with arrow and then registered with duckdb, the querying is happening all inside of duckdb?
