### 2022-02-17 #492 @cboettig
@jonkeane thanks so much for joining us on this!  Yes, I mean to test on the nightly, as you see @Pakillo is testing against the current 7.0 release.  `gbifdb` does not use `to_arrow()` anywhere though.  `gbifdb` uses `arrow::to_duckdb` (optionally) on the remote data source, which seems to result in crashes (though it seems okay if `to_duckdb()` is used on local parquet instead of over the S3 interface).

`gbifdb` also uses `duckdb::duckdb_register_arrow` on an `arrow::open_dataset()` called on the local parquet files, which seems to cause trouble as well. This seems to be nearly the same thing as calling `arrow::to_duckdb()` (a quick skim of source code in both cases is that both are creating a "View" in duckdb of the parquet tables, right? so I think either case is handing off to `duckdb` for the actual queries there-after), so I'm not sure why the one (`arrow::to_duckdb`) works while the other fails.  (Though if that's true, it's not clear to me how `arrow::to_duckdb()` is working on remote sources in the first place?)
