### 2022-02-17 #492 @jhollist
Just wanted to let you all know that I haven't been ignoring this, just didn't have anything useful to add.  At this point it sounds like we are making progress on tracking down the source of the crashes, but still trying to figure out the solution.  If there is anything you need from me, just let me know and I will keep an eye on the conversation.
