### 2022-02-17 #492 @Pakillo
Hi,

I'm happy to help fix this, but I don't know what is needed exactly (beyond what I've shared above). How can I generate a stack trace of those problems?

Also, I wonder: am I the only one getting these crashes? Why? Do you have much larger RAM available?
