### 2022-02-18 #492 @jhollist
@Pakillo

```
gbif <- gbif_remote(to_duckdb = TRUE)
gbif %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year)
```
Does not run on my machine either.  I don't ever get to a crash, it hangs.  I let it run for several hours last night  just to be sure it wasn't just taking long.

@cboettig Do you have a plan for solving these or do we want to put the submission on hold until some of the dependency issues clear up?
