### 2022-02-18 #492 @cboettig
Hi all,

I really appreciate all the help and testing!  Just wanted to make sure I'm keeping straight what is and isn't crashing for everyone.

## Remote sources

`gbifdb` has two ways to access data without download:
1. `gbif_remote(to_duckdb = FALSE)`
2. `gbif_remote(to_duckdb = TRUE)`

Am I correct that everyone sees issues with the second case but not the first on current releases, `arrow` 7.0 and duckdb 0.3.2?

## Local

`gbifdb` has two ways to do local parquet access currently (1-2 below), but we're also now discussing 2 additional strategies (3-4) by which arrow could access the local dataset:

1. `gbif_local(backend = "arrow")` (which is `arrow::open_dataset()` + `duckdb::duckdb_register_arrow`)
2. `gbif_local(backend = "duckdb")` (which is "pure duckdb")
3. `arrow::open_dataset()` + `to_duckdb()`
4. `arrow::open_dataset()` without `to_duckdb()`

Do I have it correct that: 1 & 3 crash for everyone, but 2 and 4 work for everyone?  (I'm double-checking these on my end too with dev versions of the packages)>

I feel like local access strategies 2 & 3 ought to be practically identical, so I'm a bit puzzled there.

Practically speaking, if my summary is correct, I can make is that for `gbif_remote()`, option 1 is the default, and for `gbif_local()`, that option 2 is the default.  (this makes the remote option "pure arrow" and the local option "pure duckdb", with the minor downside that "pure arrow" doesn't return a tibble or support windowing, but can live with that until `to_duckdb` is fixed).

I don't want to over-impose on our rOpenSci reviewers & editors here, but really appreciate everyone's perspectives and input in weighing these alternatives.

For the cases that crash, I definitely want to make sure we can get detailed, reproducible bug reports to arrow JIRA and duckdb GitHub trackers.
