### 2022-02-18 #492 @stephhazlitt
@cboettig I am working through the various strategies today and will report back with details.
