### 2022-02-19 #492 @stephhazlitt
I was not experiencing the same crashing as others, however as a confirmation, I thought I would try the various paths out again. I installed the `patch-download` version before running these.

### Remote Options

#### 1. arrow::open_dataset()

```
library(gbifdb)
library(dplyr)
library(arrow)

> ## arrow::open_dataset()
> bucket <- s3_bucket("gbif-open-data-us-east-1/occurrence/2022-02-01/occurrence.parquet")
> gbif_bucket <- open_dataset(bucket)
>
> system.time({
+ df_arrow <- gbif_bucket |>
+   filter(phylum == "Chordata", year > 2010) |>
+   count(class, year) |>
+   collect()
+ df_arrow
+ })
    user   system  elapsed
 182.430   35.228 1008.520
```

#### 2. gbifdb::gbif_remote(to_duckdb = FALSE)

```
> ## gbif::gbif_remote w/ arrow
> gbif_arrow <- gbif_remote(bucket = "gbif-open-data-us-east-1",
+                           to_duckdb = FALSE)
>
> system.time({
+ df_gbif_arrow <- gbif_arrow |>
+   filter(phylum == "Chordata", year > 2010) |>
+   count(class, year) |>
+   collect()
+ df_gbif_arrow
+ })
* Closing connection 0
    user   system  elapsed
 184.142   36.059 1055.453
```

#### 3. gbifdb::gbif_remote(to_duckdb = TRUE)

```
> gbif_duckdb <- gbif_remote(bucket = "gbif-open-data-us-east-1",
+                         to_duckdb = TRUE)
* Closing connection 0
>
> system.time({
+ df_gbif_duckdb <- gbif_duckdb |>
+   filter(phylum == "Chordata", year > 2010) |>
+   count(class, year) |>
+   collect()
+ df_gbif_duckdb
+ })
    user   system  elapsed
 283.422   82.892 2137.000
```

#### Session Info

```
> session_info()
─ Session info ─────────────────────────────────────
 setting  value
 version  R version 4.1.2 (2021-11-01)
 os       macOS Big Sur 11.6
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       America/Vancouver
 date     2022-02-18
 rstudio  2021.09.2+382 Ghost Orchid (desktop)
 pandoc   NA

─ Packages ─────────────────────────────────────────
 package     * version date (UTC) lib source
 arrow       * 7.0.0   2022-02-10 [1] CRAN (R 4.1.1)
 assertthat    0.2.1   2019-03-21 [1] CRAN (R 4.1.0)
 bit           4.0.4   2020-08-04 [1] CRAN (R 4.1.1)
 bit64         4.0.5   2020-08-30 [1] CRAN (R 4.1.0)
 blob          1.2.2   2021-07-23 [1] CRAN (R 4.1.0)
 brio          1.1.3   2021-11-30 [1] CRAN (R 4.1.1)
 cachem        1.0.6   2021-08-19 [1] CRAN (R 4.1.1)
 callr         3.7.0   2021-04-20 [1] CRAN (R 4.1.0)
 cli           3.2.0   2022-02-14 [1] CRAN (R 4.1.2)
 crayon        1.5.0   2022-02-14 [1] CRAN (R 4.1.2)
 DBI           1.1.2   2021-12-20 [1] CRAN (R 4.1.2)
 dbplyr        2.1.1   2021-04-06 [1] CRAN (R 4.1.0)
 desc          1.4.0   2021-09-28 [1] CRAN (R 4.1.1)
 devtools    * 2.4.3   2021-11-30 [1] CRAN (R 4.1.1)
 dplyr       * 1.0.8   2022-02-08 [1] CRAN (R 4.1.2)
 duckdb        0.3.2   2022-02-07 [1] CRAN (R 4.1.2)
 ellipsis      0.3.2   2021-04-29 [1] CRAN (R 4.1.0)
 fansi         1.0.2   2022-01-14 [1] CRAN (R 4.1.1)
 fastmap       1.1.0   2021-01-25 [1] CRAN (R 4.1.0)
 fs            1.5.2   2021-12-08 [1] CRAN (R 4.1.2)
 gbifdb      * 0.1.1   2022-02-18 [1] Github (cboettig/gbifdb@51c6d30)
 generics      0.1.2   2022-01-31 [1] CRAN (R 4.1.2)
 glue          1.6.1   2022-01-22 [1] CRAN (R 4.1.1)
 lifecycle     1.0.1   2021-09-24 [1] CRAN (R 4.1.1)
 magrittr      2.0.2   2022-01-26 [1] CRAN (R 4.1.1)
 memoise       2.0.1   2021-11-26 [1] CRAN (R 4.1.1)
 pillar        1.7.0   2022-02-01 [1] CRAN (R 4.1.2)
 pkgbuild      1.3.1   2021-12-20 [1] CRAN (R 4.1.2)
 pkgconfig     2.0.3   2019-09-22 [1] CRAN (R 4.1.0)
 pkgload       1.2.4   2021-11-30 [1] CRAN (R 4.1.1)
 praise        1.0.0   2015-08-11 [1] CRAN (R 4.1.0)
 prettyunits   1.1.1   2020-01-24 [1] CRAN (R 4.1.0)
 processx      3.5.2   2021-04-30 [1] CRAN (R 4.1.0)
 ps            1.6.0   2021-02-28 [1] CRAN (R 4.1.0)
 purrr         0.3.4   2020-04-17 [1] CRAN (R 4.1.0)
 R6            2.5.1   2021-08-19 [1] CRAN (R 4.1.1)
 remotes       2.4.2   2021-11-30 [1] CRAN (R 4.1.1)
 rlang         1.0.1   2022-02-03 [1] CRAN (R 4.1.2)
 rprojroot     2.0.2   2020-11-15 [1] CRAN (R 4.1.0)
 sessioninfo   1.2.2   2021-12-06 [1] CRAN (R 4.1.1)
 testthat    * 3.1.2   2022-01-20 [1] CRAN (R 4.1.1)
 tibble        3.1.6   2021-11-07 [1] CRAN (R 4.1.1)
 tidyselect    1.1.1   2021-04-30 [1] CRAN (R 4.1.0)
 usethis     * 2.1.5   2021-12-09 [1] CRAN (R 4.1.1)
 utf8          1.2.2   2021-07-24 [1] CRAN (R 4.1.0)
 vctrs         0.3.8   2021-04-29 [1] CRAN (R 4.1.0)
 withr         2.4.3   2021-11-30 [1] CRAN (R 4.1.1)

 [1] /Users/stephhazlitt/Rlibrary
 [2] /Library/Frameworks/R.framework/Versions/4.1-arm64/Resources/library
```
