### 2022-02-19 #492 @Pakillo
Hi, here are my results

# Remote options

## 1. arrow::open_dataset()

```r
library(gbifdb)
library(dplyr)
library(arrow)

## arrow::open_dataset()
bucket <- s3_bucket("gbif-open-data-eu-central-1/occurrence/2022-02-01/occurrence.parquet")
gbif_bucket <- open_dataset(bucket)

system.time({
  df_arrow <- gbif_bucket |>
    filter(phylum == "Chordata", year > 2010) |>
    count(class, year) |>
    collect()
  df_arrow
})
#>    user  system elapsed
#> 235.360  27.849 203.743

## 2. gbifdb::gbif_remote(to_duckdb = FALSE)

``` r
library(gbifdb)
library(dplyr)

gbif_arrow <- gbif_remote(bucket = "gbif-open-data-eu-central-1",
                          to_duckdb = FALSE)

system.time({
  df_gbif_arrow <- gbif_arrow |>
    filter(phylum == "Chordata", year > 2010) |>
    count(class, year) |>
    collect()
  df_gbif_arrow
})
#>    user  system elapsed
#> 235.260  26.648 181.525
```

## 3. gbifdb::gbif_remote(to_duckdb = TRUE)

CRASHES R (due to RAM)
