### 2022-02-20 #492 @Pakillo
# Local database

## 1. gbif_local(backend = "arrow") - CRASH

(which is arrow::open_dataset() + duckdb::duckdb_register_arrow)

```r
gbif <- gbif_local(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/",
                     backend = "arrow")
gbif %>%
  filter(phylum == "Chordata", year > 2010) %>%
  count(class, year) %>%
  collect()
```

CRASH

## 2. gbif_local(backend = "duckdb") - WORKS

(which is "pure duckdb")

``` r
system.time({
  gbif <- gbif_local(dir = "/media/frs/Elements/gbifdb/occurrence.parquet/",
                     backend = "duckdb")
  gbif %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year) %>%
    collect()
})
#>    user  system elapsed
#>  78.476   2.166  94.556
```

## 3. arrow::open_dataset() + to_duckdb() - CRASH

```r
system.time({
  arrow::open_dataset("/media/frs/Elements/gbifdb/occurrence.parquet/") %>%
  arrow::to_duckdb() %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year) %>%
    collect()
})
```

CRASHES R

## 4. arrow::open_dataset() without to_duckdb() - WORKS

``` r
system.time({
  arrow::open_dataset("/media/frs/Elements/gbifdb/occurrence.parquet/") %>%
    filter(phylum == "Chordata", year > 2010) %>%
    count(class, year) %>%
    collect()
})
#>    user  system elapsed
#> 273.518  15.553  40.928
```

# Conclusions

- gbif_remote works well when to_duckdb = FALSE.

- gbif_local works well when backend = "duckdb"

So, the problem seems to be when converting arrow to duckdb
