### 2022-02-20 #492 @stephhazlitt
#### Local Options

```
library(gbifdb)
library(dplyr)
library(arrow)

gbif_dir <- "/Users/stephhazlitt/dev/gbif-data/occurrence.parquet/"
```

**1. gbif_local(backend = "arrow") (which is arrow::open_dataset() + duckdb::duckdb_register_arrow)**

**WORKS**

```
# 1. gbif_local(backend = "arrow") (which is arrow::open_dataset() + duckdb::duckdb_register_arrow)
gbif_local(dir = gbif_dir,
           backend = "arrow") |>
  filter(phylum == "Chordata", year > 2010) |>
  count(class, year) |>
  collect()
#> # A tibble: 143 × 3
#> # Groups:   class [15]
#>    class           year     n
#>    <chr>          <int> <dbl>
#>  1 Reptilia        2018 10837
#>  2 Actinopterygii  2019 21428
#>  3 Reptilia        2013  6236
#>  4 Actinopterygii  2014 70435
#>  5 Actinopterygii  2013 91919
#>  6 <NA>            2013    60
#>  7 Reptilia        2014  6128
#>  8 Amphibia        2017  9118
#>  9 Actinopterygii  2016 58846
#> 10 Reptilia        2017  9534
#> # … with 133 more rows
```

**2. gbif_local(backend = "duckdb") (which is "pure duckdb")**

**WORKS**

```
# 2. gbif_local(backend = "duckdb") (which is "pure duckdb")
gbif_local(dir = gbif_dir,
           backend = "duckdb") |>
  filter(phylum == "Chordata", year > 2010) |>
  count(class, year) |>
  collect()
#> # A tibble: 143 × 3
#> # Groups:   class [15]
#>    class       year     n
#>    <chr>      <int> <dbl>
#>  1 Ascidiacea  2018   187
#>  2 Ascidiacea  2011   302
#>  3 Ascidiacea  2012   276
#>  4 Ascidiacea  2014   757
#>  5 Ascidiacea  2016   335
#>  6 Ascidiacea  2015   392
#>  7 Ascidiacea  2013   169
#>  8 Ascidiacea  2017   394
#>  9 Ascidiacea  2019   161
#> 10 Ascidiacea  2020   144
#> # … with 133 more rows
```

**3. arrow::open_dataset() + to_duckdb()**

**CRASHES**

```
# 3. arrow::open_dataset() + to_duckdb()
arrow::open_dataset(gbif_dir) |>
  to_duckdb() |>
  filter(phylum == "Chordata", year > 2010) |>
  count(class, year) |>
  collect()
```

**4. arrow::open_dataset() without to_duckdb()**

**WORKS**

```
# 4. arrow::open_dataset() without to_duckdb()
arrow::open_dataset(gbif_dir) |>
  filter(phylum == "Chordata", year > 2010) |>
  count(class, year) |>
  collect()
#> # A tibble: 143 × 3
#> # Groups:   class [15]
#>    class           year       n
#>    <chr>          <int>   <int>
#>  1 Aves            2015 1880023
#>  2 Reptilia        2012    5686
#>  3 Reptilia        2015    6219
#>  4 Reptilia        2013    6236
#>  5 Reptilia        2017    9534
#>  6 Mammalia        2018   45356
#>  7 Mammalia        2013   57009
#>  8 Actinopterygii  2012  105466
#>  9 Reptilia        2011    5481
#> 10 Mammalia        2019   58039
#> # … with 133 more rows
```
