### 2022-02-20 #492 @stephhazlitt
I am not sure why now, as it often just crashed out right, but when I was running the `3. arrow::open_dataset() + to_duckdb()` my computer prompted me to force close applications as it was running out of memory -- so I guess it is a RAM issue. No issue with any of the options with the very small example data set:

``` r
library(gbifdb)
library(dplyr)
library(arrow)

# 1. gbif_local(backend = "arrow") (which is arrow::open_dataset() + duckdb::duckdb_register_arrow)
gbif_local(dir = gbif_example_data(),
           backend = "arrow") |>
  filter(kingdom == "Plantae") |>
  group_by(class) |>
  count() |>
  collect()
#> # A tibble: 6 × 2
#>   class              n
#>   <chr>          <dbl>
#> 1 Lycopodiopsida    35
#> 2 Liliopsida       461
#> 3 Magnoliopsida    317
#> 4 Polypodiopsida   135
#> 5 Pinopsida         27
#> 6 Gnetopsida         7

# 2. gbif_local(backend = "duckdb") (which is "pure duckdb")
gbif_local(dir = gbif_example_data(),
           backend = "duckdb") |>
  filter(kingdom == "Plantae") |>
  group_by(class) |>
  count() |>
  collect()
#> # A tibble: 6 × 2
#>   class              n
#>   <chr>          <dbl>
#> 1 Lycopodiopsida    35
#> 2 Liliopsida       461
#> 3 Magnoliopsida    317
#> 4 Polypodiopsida   135
#> 5 Pinopsida         27
#> 6 Gnetopsida         7

# 3. arrow::open_dataset() + to_duckdb()
arrow::open_dataset(gbif_example_data()) |>
  to_duckdb() |>
  filter(kingdom == "Plantae") |>
  group_by(class) |>
  count() |>
  collect()
#> # A tibble: 6 × 2
#>   class              n
#>   <chr>          <dbl>
#> 1 Lycopodiopsida    35
#> 2 Liliopsida       461
#> 3 Magnoliopsida    317
#> 4 Polypodiopsida   135
#> 5 Pinopsida         27
#> 6 Gnetopsida         7

# 4. arrow::open_dataset() without to_duckdb()
arrow::open_dataset(gbif_example_data()) |>
  filter(kingdom == "Plantae") |>
  group_by(class) |>
  count() |>
  collect()
#> # A tibble: 6 × 2
#> # Groups:   class [6]
#>   class              n
#>   <chr>          <int>
#> 1 Lycopodiopsida    35
#> 2 Liliopsida       461
#> 3 Magnoliopsida    317
#> 4 Polypodiopsida   135
#> 5 Pinopsida         27
#> 6 Gnetopsida         7
```
