### 2022-03-09 #492 @cboettig
@jhollist thanks for the ping!

Good question, yeah, can't really predict how long it might be to see the `to_duckdb()` or `duckdb_register_arrow()` mechanisms fixed.

It sounds like we could just make `gbif_local()` be "pure-duckdb" and `gbif_remote()` be pure `arrow` and we would not have any issues. Once those mechanisms stabilize, we could put them back in (I think they have the potential for nicer consistency and performance once they are finally resolved, but kinda a minor thing).  Does that sound reasonable to everyone?
