### 2022-03-09 #492 @jhollist
That would work for me.  @stephhazlitt  and @Pakillo does @cboettig suggestion work for you.
