### 2022-03-09 af80552
pure arrow / pure duckdb
