### 2022-03-14 1d16c36
configurable defaults
Primarily, this enables alternate remote endpoints without altering argument names.
