+++ man/gbif_version.Rd
@@ -26,9 +26,15 @@ gbif_version(
 \description{
 Can also return latest locally downloaded version, or list all versions
 }
+\details{
+A default version can be set using option \code{gbif_default_version}
+}
 \examples{
 ## Latest local version available:
 gbif_version(local=TRUE)
+## default version
+options(gbif_default_version="2021-01-01")))
+gbif_version()
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 ## Latest online version available:
 gbif_version()
