+++ README.Rmd
@@ -14,8 +14,12 @@ knitr::opts_chunk$set(
   out.width = "100%"
 )
 
-
-Sys.setenv("GBIF_HOME"="/home/shared-data/gbif")
+## Use local gbif copy
+options(gbif_unset_aws=FALSE, 
+        gbif_default_bucket="gbif",
+        gbif_default_version="2021-11-01")
+Sys.setenv("GBIF_HOME"="/home/shared-data/gbif",
+           AWS_S3_ENDPOINT = "minio.thelio.carlboettiger.info")
 
 ```
 
@@ -84,7 +88,8 @@ We can now perform most `dplyr` operations:
 ```{r}
 gbif %>%
   filter(phylum == "Chordata", year > 1990) %>%
-  count(class, year)
+  count(class, year) %>%
+  collect()
 ```
 
 By default, this relies on an `arrow` connection, which currently lacks support for some more complex windowed operations in `dplyr`.
@@ -111,7 +116,7 @@ This requires over 100 GB free disk space, and will be a time-consuming process 
 However, once downloaded, future queries will run much much faster, particularly if you are network-limited.
 Users can download the current release of GBIF to local storage like so:
 
-```r
+```{r}
 gbif_download()
 ```
 
@@ -120,7 +125,7 @@ An alternative directory can be provided by setting the environmental variable, 
 or providing the path to the directory containing the parquet files directly.
 
 Once you have downloaded the parquet-formatted GBIF data, 
-`gbif_local()` will establish a connection to thes local parquet files. 
+`gbif_local()` will establish a connection to these local parquet files. 
 
 ```{r}  
 gbif <- gbif_local()
@@ -153,7 +158,7 @@ growth %>%
  collect() %>%
   mutate(class = fct_lump_n(class, 6)) %>%
   ggplot(aes(year, n, fill=class)) + geom_col() +
-  ggtitle("GBIF observations by class")
+  ggtitle("GBIF observations of vertebrates by class")
 ```
 
 
