### 2022-03-15 #492 @cboettig
Thanks!  I've switched to the duckdb-only local and arrow-only remote as the defaults, https://github.com/cboettig/gbifdb/pull/6
