### 2022-03-18 #492 @jhollist
@cboettig Thanks for those changes.  I can confirm that gbif_remote() works for me locally.  I didn't test gbif_local() as I don't have  100GB free!  That being said, looks like no one had issues with the pure duckdb versions.

Let me refresh my memory on next steps here, but I think we are very close!

@Pakillo and @stephhazlitt any remaining concerns?
