### 2022-03-22 #492 @jhollist
@cboettig we are good to go!

@Pakillo and @stephhazlitt I went ahead and checked off the two remaining items on your checklists.  The subsequent discussion indicates you are both good to go with accepting `gbifdb`.

Congrats, Carl!  Another great contribution.
