### 2022-03-22 #492 @ropensci-review-bot
Approved! Thanks @cboettig for submitting and @Pakillo, @stephhazlitt for your reviews! :grin:
