### 2022-03-24 #492 @cboettig
Thanks @Pakillo .  Correct, I left those in as options but not defaults. (e.g. technically any user reading the `arrow` docs might be tempted to try `arrow::to_duckdb()` on the `gbif_remote()` connection even it weren't a built-in option, and they would experience similar crashes).  I'm hoping having the options in there remind me to revisit these issues as duckdb and arrow continue to evolve, and maybe even draws some more attention to these bugs from the upstream developers!
